requirements.txt
py-cord==1.7.3
yt-dlp==2021.12.27
asyncio==3.4.3
python-dotenv==0.19.0
